Add reusable Quiz component
